Using `IHttpClientFactory` to create `HttpClient`s
This is a follow up-PR to #2937, #2953 and #2962 because in the original PR (#2937) there was a missing `context.Services.AddHttpClient()` inside the module - which is now part of the module in the `dev` branch.
I haven't used named clients because currently there won't be any benefit from my point of view since the `HttpClient` only gets used to fire the `IdentityModel` requests using extension methods provided by the `IdentityModel` package.