Used different names for `ReadOnlyRepository` and `Repository` in application services
Still using `protected new` on `CrudAppService` because the repository member in `AbstractKeyCrudAppService` has to be renamed to something like `AbstractKeyRepository`, which would be a breaking change.

Renaming the repository in each descendant could lead to misunderstandings, since in `CrudAppService` descendants will be three repositories to choose from: `Repository`, `ReadOnlyRepository`, `AbstractKeyRepository`.
Currently the user must only differ between `ReadOnlyRepository` and `Repository`.

@maliming, can you review this?